Move Opentoscautils
Move the remaining methods of the quantme OpentoscaUtils to the opentosca OpentoscaUtils
