+++ components/bpmn-q/modeler-component/extensions/quantme/ui/adaptation/runtimes/AwsRuntimeHandler.js
@@ -18,7 +18,7 @@ import { startQuantmeReplacementProcess } from "../../../replacement/QuantMETran
 import {
   createNewArtifactTemplate,
   createNewServiceTemplateVersion,
-} from "../../../deployment/OpenTOSCAUtils";
+} from "../../../../opentosca/deployment/OpenTOSCAUtils";
 import {
   getInvalidModelingConstruct,
   getRequiredPrograms,
