+++ components/bpmn-q/modeler-component/extensions/quantme/ui/adaptation/runtimes/QiskitRuntimeHandler.js
@@ -18,7 +18,7 @@ import {
 import {
   createNewArtifactTemplate,
   createNewServiceTemplateVersion,
-} from "../../../deployment/OpenTOSCAUtils";
+} from "../../../../opentosca/deployment/OpenTOSCAUtils";
 import {
   getInvalidModelingConstruct,
   getRequiredPrograms,
